feat: cleanup code mapper
